Added mutations and tests

+++ IntegrationTests/Helpers/IntegrationTestBase.cs
@@ -7,36 +7,50 @@ using MongoDB.Driver;
 
 namespace IntegrationTests.Helpers;
 
-public abstract class IntegrationTestBase : TestBase, IAsyncLifetime
+[TestClass]
+public class IntegrationTestBase : TestBase
 {
-    private readonly IServiceProvider _serviceProvider;
+    private static readonly IServiceProvider ServiceProvider;
 
-    protected IntegrationTestBase()
+    static IntegrationTestBase()
     {
-        var configuration =  new ConfigurationBuilder()
+        var configuration = new ConfigurationBuilder()
             .AddJsonFile("appsettings.development.json");
-        
+
         var config = configuration.Build();
-        
-        _serviceProvider = new ServiceCollection()
+
+        ServiceProvider = new ServiceCollection()
             .AddInfra(config.GetRequiredSection("MongoDb").Get<MongoConfiguration>()!)
-            .AddOptions()
             .BuildServiceProvider();
+    }
+
+    [TestCleanup]
+    public async Task CleanUp()
+    {
+        await OperateInDocumentsFromMongoDbCollections(OperationType.DeleteMany);
+    }
+
+    [AssemblyCleanup]
+    public static Task DropAllMongoDbCollections()
+    {
+        return OperateInDocumentsFromMongoDbCollections(OperationType.DropCollection);
     }
 
     protected TService GetService<TService>()
         where TService : notnull
     {
-        return _serviceProvider.GetRequiredService<TService>();
+        return ServiceProvider.GetRequiredService<TService>();
     }
 
-    public Task InitializeAsync() => Task.CompletedTask;
-
-    public Task DisposeAsync() => DeleteAllDocumentsFromMongoDbCollections();
-
-    private async Task DeleteAllDocumentsFromMongoDbCollections()
+    private enum OperationType
     {
-        var database = _serviceProvider.GetRequiredService<IMongoDatabase>();
+        DropCollection,
+        DeleteMany
+    }
+
+    private static async Task OperateInDocumentsFromMongoDbCollections(OperationType operationType)
+    {
+        var database = ServiceProvider.GetRequiredService<IMongoDatabase>();
 
         using var cursor = await database.ListCollectionNamesAsync();
         while (await cursor.MoveNextAsync())
@@ -45,27 +59,14 @@ public abstract class IntegrationTestBase : TestBase, IAsyncLifetime
             {
                 var collection = database.GetCollection<BsonDocument>(collectionName);
 
-                // Include a brief pause between delete operations to mitigate potential issues
-                // related to eventual consistency in distributed database environments.
-                // This delay helps ensure that each delete operation has sufficient time to be 
-                // recognized across the cluster, thereby preventing race conditions or read inconsistencies
-                // that might occur in rapid succession of delete commands, especially in testing scenarios.
-                await Task.Delay(500);
-                await collection.DeleteManyAsync(Builders<BsonDocument>.Filter.Empty);
-            }
-        }
-    }
-    
-    private async Task DropAllMongoDbCollections()
-    {
-        var database = _serviceProvider.GetRequiredService<IMongoDatabase>();
-        
-        using var cursor = await database.ListCollectionNamesAsync();
-        while (await cursor.MoveNextAsync())
-        {
-            foreach (var collectionName in cursor.Current)
-            {
-                await database.DropCollectionAsync(collectionName);
+                if (operationType == OperationType.DropCollection)
+                {
+                    await database.DropCollectionAsync(collectionName);
+                }
+                else
+                {
+                    await collection.DeleteManyAsync(Builders<BsonDocument>.Filter.Empty);
+                }
             }
         }
     }
